DEPR: deprecate morphological.optimize_window
Moved _optimize_window to utils and renamed as optimize_window, so that users can directly call it. Deprecate the optimize_window function from morphological, so that it only contains baseline functions. Moved difference_matrix to utils so that users can call it as well.
